Refs 3613. Use cookies to keep track of timeouts
If a user's session times out set 'session_timeout' to true in the
browser's cookie so that new JavaScript included in layouts/_header will
pop up a modal reminding the user to login if they click back in their
browser and load an old page. This will prevent the user from submit a
form when they aren't really logged in, which is where all the
ActionController::InvalidAuthenticityToken exceptions were coming from.

Refs 3613.